+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/resources/org/eclipse/lsp4jakarta/jdt/core/messages/messages.properties
@@ -38,7 +38,7 @@ ChangeReturnTypeToVoid = Change return type to void
 # RemoveDynamicConstraintAnnotationQuickFix
 RemoveConstraintAnnotation = Remove constraint annotation {0} from element
 
-# BeanValidationDiagnosticsParticipant
+# ValidationDiagnosticsParticipant
 ConstraintAnnotationsMethod = Constraint annotations are not allowed on static methods.
 ConstraintAnnotationsField = Constraint annotations are not allowed on static fields.
 AnnotationBooleanMethods = The {0} annotation can only be used on boolean and Boolean type methods.
