+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/projects/demo-servlet-no-diagnostics/src/main/java/io/openliberty/sample/jakarta/validation/FieldConstraintValidation.java
@@ -1,9 +1,9 @@
-package io.openliberty.sample.jakarta.beanvalidation;
+package io.openliberty.sample.jakarta.validation;
 
 import java.util.Calendar;
 import java.util.List;
 
-import jakarta.validation.constraints.*;
+import jakarta.fake.constraints.*;
 
 public class FieldConstraintValidation {
