+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/projects/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/validation/FieldConstraintValidation.java
@@ -1,9 +1,9 @@
-package io.openliberty.sample.jakarta.beanvalidation;
+package io.openliberty.sample.jakarta.validation;
 
 import java.util.Calendar;
 import java.util.List;
 
-import jakarta.fake.constraints.*;
+import jakarta.validation.constraints.*;
 
 public class FieldConstraintValidation {
