+++ jakarta.ls/src/main/resources/validation.json
